// src/types/page.ts
export interface PageConfig {
    sections: {
        jobs: {
            title: string;
            viewMoreText: string;
            viewMoreLink: string;
        };
        projects: {
            title: string;
            viewMoreText: string;
            viewMoreLink: string;
        };
        certifications: {
            title: string;
            viewMoreText: string;
            viewMoreLink: string;
        };
    };
}